Allow multiple separate parsers
A global parser is always defined (so the API does not change) but it is now
allowed to build multiple separate parsers with the cliargs.new function.

There is still a problem about the module wiping itself after calling
parse() unless _G._TEST is defined... don't know how to solve it yet.